split Explore Map apart
